fix: Bazel renamed the modquery command to mod (#4937)
Fixes https://github.com/aspect-build/aspect-cli/issues/555

https://github.com/bazelbuild/bazel/pull/18921/files#diff-18cce070c24a0b6eaf8de41353b962efebb9d857dda06c044032ae1a50c3e79b
is where the rename happened.

Note, unlike other commands, I did not attempt to reproduce all the
documentation.
The docs on the Bazel site are pretty good since this command is newly
added.

---

### Type of change

- Bug fix (change which fixes an issue)

### Test plan

- Covered by existing test cases

GitOrigin-RevId: 0884f089bebac1377c46e132b6ce38cd06a2a41d
