+++ cmd/aspect/mod/BUILD.bazel
@@ -1,12 +1,12 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "modquery",
-    srcs = ["modquery.go"],
-    importpath = "aspect.build/cli/cmd/aspect/modquery",
+    name = "mod",
+    srcs = ["mod.go"],
+    importpath = "aspect.build/cli/cmd/aspect/mod",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/aspect/modquery",
+        "//pkg/aspect/mod",
         "//pkg/aspect/root/flags",
         "//pkg/bazel",
         "//pkg/interceptors",
