+++ cmd/aspect/mod/mod.go
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2022 Aspect Build Systems, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package mod
+
+import (
+	"github.com/spf13/cobra"
+
+	"aspect.build/cli/pkg/aspect/mod"
+	"aspect.build/cli/pkg/aspect/root/flags"
+	"aspect.build/cli/pkg/bazel"
+	"aspect.build/cli/pkg/interceptors"
+	"aspect.build/cli/pkg/ioutils"
+)
+
+func NewDefaultCmd() *cobra.Command {
+	return NewCmd(ioutils.DefaultStreams, bazel.WorkspaceFromWd)
+}
+
+func NewCmd(streams ioutils.Streams, bzl bazel.Bazel) *cobra.Command {
+	cmd := &cobra.Command{
+		Use:     "mod <subcommand> [<options>] [<args> ...]",
+		Short:   "Tools to work with the bzlmod external dependency graph",
+		Long:    `See https://bazel.build/external/mod-command`,
+		GroupID: "built-in",
+		RunE: interceptors.Run(
+			[]interceptors.Interceptor{
+				flags.FlagsInterceptor(streams),
+			},
+			mod.New(streams, bzl).Run,
+		),
+	}
+
+	return cmd
+}
