+++ cmd/aspect/root/BUILD.bazel
@@ -25,7 +25,7 @@ go_library(
         "//cmd/aspect/license",
         "//cmd/aspect/lint",
         "//cmd/aspect/mobileinstall",
-        "//cmd/aspect/modquery",
+        "//cmd/aspect/mod",
         "//cmd/aspect/outputs",
         "//cmd/aspect/print",
         "//cmd/aspect/printaction",
