+++ pkg/aspect/mod/BUILD.bazel
@@ -1,9 +1,9 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "modquery",
-    srcs = ["modquery.go"],
-    importpath = "aspect.build/cli/pkg/aspect/modquery",
+    name = "mod",
+    srcs = ["mod.go"],
+    importpath = "aspect.build/cli/pkg/aspect/mod",
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/bazel",
